Added proper life cycle management for standalone server. Upgraded Jersey version to 2.4. Added further tests.

+++ pdf-converter-local-demo/src/main/java/no/kantega/pdf/demo/LinkPanel.java
@@ -36,7 +36,7 @@ public class LinkPanel extends Panel {
             } else if (size < 1024L * 1024L) {
                 return String.format("~%d kB", size / 1024L);
             } else {
-                double truncatedMegabyte = ((double) (size / 1024L)) / 1024L;
+                double truncatedMegabyte = (size / 1024L) / 1024d;
                 return String.format("~%.2f MB", truncatedMegabyte);
             }
         }
